+++ R/explore.R
@@ -1,3 +1,7 @@
+## default color "#ADD8E6", #6BAED6 (original)
+## default color "#ADD8E6", "#6BAED6" (test)
+## default color "#ADD8E6", "#7BB8DA" (actual)
+
 #' Explore categorical variable + target
 #'
 #' Create a plot to explore relation between categorical variable and a binary target
@@ -278,7 +282,7 @@ target_explore_num <- function(data, var, target = "target_ind", min_val = NA, m
 #' @return Plot object (bar chart)
 #' @export
 
-explore_bar <- function(data, var, target, flip = NA, title = "", numeric = NA, max_cat = 30, max_target_cat = 5, color = c("lightgrey", "#939FB9"), legend_position = "right", label, label_size = 2.7, ...)  {
+explore_bar <- function(data, var, target, flip = NA, title = "", numeric = NA, max_cat = 30, max_target_cat = 5, color = c("#ADD8E6", "#7BB8DA"), legend_position = "right", label, label_size = 2.7, ...)  {
 
   # define variables for CRAN-package check
   na_ind <- NULL
@@ -538,7 +542,7 @@ explore_bar <- function(data, var, target, flip = NA, title = "", numeric = NA, 
 #' explore_density(iris, Sepal.Length, target = is_virginica)
 #' @export
 
-explore_density <- function(data, var, target, title = "", min_val = NA, max_val = NA, color = c("lightgrey", "#939FB9"), auto_scale = TRUE, max_target_cat = 5, ...)   {
+explore_density <- function(data, var, target, title = "", min_val = NA, max_val = NA, color = c("#ADD8E6", "#7BB8DA"), auto_scale = TRUE, max_target_cat = 5, ...)   {
 
   # check parameter data
   check_data_frame_non_empty(data)
@@ -629,8 +633,9 @@ explore_density <- function(data, var, target, title = "", min_val = NA, max_val
     if (show_mean_var) {
       if ((auto_scale == FALSE) | (mean_var <= max_val)) {
         p <- p + geom_vline(xintercept = mean_var,
-                            color = "#7f7f7f", alpha = 0.5,
-                            linetype = "dashed", lwd = 1)
+                            #color = "#7f7f7f", alpha = 0.5,
+                            color = mix_color(color[1], "black", n = 5)[2],
+                            alpha = 0.75, linetype = "dashed", lwd = 1)
       }
     }
 
@@ -714,7 +719,7 @@ explore_density <- function(data, var, target, title = "", min_val = NA, max_val
 #' explore_all(iris, target = is_virginica)
 #' @export
 
-explore_all <- function(data, n, target, ncol = 2, targetpct, color = c("lightgrey", "#939FB9"), split = TRUE)  {
+explore_all <- function(data, n, target, ncol = 2, targetpct, color = c("#ADD8E6", "#7BB8DA"), split = TRUE)  {
 
   # check parameter data
   check_data_frame_non_empty(data)
@@ -859,7 +864,7 @@ explore_all <- function(data, n, target, ncol = 2, targetpct, color = c("lightgr
 #' explore_cor(iris, x = Sepal.Length, y = Sepal.Width)
 #' @export
 
-explore_cor <- function(data, x, y, target, bins = 8, min_val = NA, max_val = NA, auto_scale = TRUE, title = NA, color = "grey", ...)  {
+explore_cor <- function(data, x, y, target, bins = 8, min_val = NA, max_val = NA, auto_scale = TRUE, title = NA, color = c("#ADD8E6", "#7BB8DA"), ...)  {
 
   # check parameter data
   check_data_frame_non_empty(data)
@@ -960,7 +965,7 @@ explore_cor <- function(data, x, y, target, bins = 8, min_val = NA, max_val = NA
     ## points x,y + color by target
     p <- data %>%
       ggplot(aes(x = !!x_quo, y = !!y_quo, color = !!target_quo)) +
-      geom_point(alpha = 0.60, size = 2.5) +
+      geom_point(alpha = 0.75, size = 2.5) +
       theme(
         panel.background = element_rect("white"),
         panel.grid.major = element_line("grey85"),
@@ -976,12 +981,15 @@ explore_cor <- function(data, x, y, target, bins = 8, min_val = NA, max_val = NA
     } else {
 
       ## points x,y
+      color_dark <- mix_color(color[1], "black", n = 5)[2]
       p <- data %>%
         ggplot(aes(x = !!x_quo, y = !!y_quo)) +
-        geom_point(alpha = 0.6, size = 2.5, color = color[1]) +
+        geom_point(alpha = 0.6, size = 2.5, shape = 21,
+                   color = color_dark, fill = color[1]) +
         geom_abline(intercept = reg_intercept, slope = reg_slope,
-                    color = "#7f7f7f", alpha = 0.5,
-                    linetype = "solid", size = 1) +
+                    #color = "#7f7f7f", alpha = 0.5,
+                    color = color_dark,
+                    alpha = 0.75, linetype = "solid", size = 1) +
         theme(
           panel.background = element_rect("white"),
           panel.grid.major = element_line("grey85"),
@@ -1186,7 +1194,7 @@ explore_tbl <- function(data, n)  {
 #' }
 #' @export
 
-explore_shiny <- function(data, target, color = c("lightgrey", "#939FB9"))  {
+explore_shiny <- function(data, target, color = c("#ADD8E6", "#7BB8DA"))  {
 
   # check if interactive session
   if (!interactive()) {
@@ -1731,7 +1739,7 @@ explore_targetpct <- function(data, var, target = NULL, title = NA, min_val = NA
 #'   explore_count(Species)
 #' @export
 
-explore_count <- function(data, cat, n, target, pct = FALSE, split = TRUE, title = NA, numeric = FALSE, max_cat = 30, max_target_cat = 5, color = c("lightgrey", "#939FB9"), flip = NA)  {
+explore_count <- function(data, cat, n, target, pct = FALSE, split = TRUE, title = NA, numeric = FALSE, max_cat = 30, max_target_cat = 5, color = c("#ADD8E6", "#7BB8DA"), flip = NA)  {
 
   # define variables for CRAN-package check
   plot_cat <- NULL