+++ NEWS.md
@@ -1,3 +1,15 @@
+# explore 1.3.2
+
+## New features
+
+* change default color from grey to "#ADD8E6", "#7BB8DA"
+* add adaptive border color to geom_point plots (correlation x,y)
+* new adaptive color for mean & regression line
+
+## Bug fixes / internal
+
+* drop CRAN download statistics
+
 # explore 1.3.1
 
 ## New features